hf-space/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.8.0+cpu
torchaudio==2.8.0+cpu
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
coqui-tts==0.27.0
transformers==4.57.6
python-multipart==0.0.20
soundfile==0.12.1
